+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `BodyState` contain self id in meta.
 - `MindState` contain self id in meta.
+- `GameStateHelper.addActress` return body and mind too.
 
 ### Fixed